Refactor port mapping logic to helpers
