+++ pkg/disruptors/service.go
@@ -6,9 +6,7 @@ import (
 	"time"
 
 	"github.com/grafana/xk6-disruptor/pkg/kubernetes"
-
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"github.com/grafana/xk6-disruptor/pkg/kubernetes/helpers"
 )
 
 // ServiceDisruptor defines operations for injecting faults in services
@@ -27,13 +25,10 @@ type ServiceDisruptorOptions struct {
 // serviceDisruptor is an instance of a ServiceDisruptor
 type serviceDisruptor struct {
 	ctx        context.Context
-	name       string
+	service    string
 	namespace  string
 	options    ServiceDisruptorOptions
-	service    *corev1.Service
-	endpoints  *corev1.Endpoints
-	k8s        kubernetes.Kubernetes
-	mapper     PortMapper
+	helper     helpers.ServiceHelper
 	controller AgentController
 }
 
@@ -41,31 +36,24 @@ type serviceDisruptor struct {
 func NewServiceDisruptor(
 	ctx context.Context,
 	k8s kubernetes.Kubernetes,
-	name string,
+	service string,
 	namespace string,
 	options ServiceDisruptorOptions,
 ) (ServiceDisruptor, error) {
-	if name == "" {
+	if service == "" {
 		return nil, fmt.Errorf("must specify a service name")
 	}
-	service, err := k8s.CoreV1().
-		Services(namespace).
-		Get(ctx, name, metav1.GetOptions{})
+
+	helper := k8s.NamespacedHelpers(namespace)
+
+	targets, err := helper.GetTargets(ctx, service)
 	if err != nil {
-		return nil, fmt.Errorf("failed to retrieve target service %s: %w", service, err)
+		return nil, err
 	}
 
-	ep, err := k8s.CoreV1().Endpoints(namespace).Get(ctx, name, metav1.GetOptions{})
-	if err != nil {
-		return nil, fmt.Errorf("failed to retrieve endpoints for service %s: %w", service, err)
-	}
-
-	mapper := NewPortMapper(ctx, service, ep)
-
-	targets := getTargetPods(ep)
 	controller := NewAgentController(
 		ctx,
-		k8s.NamespacedHelpers(namespace),
+		helper,
 		namespace,
 		targets,
 		options.InjectTimeout,
@@ -78,95 +66,11 @@ func NewServiceDisruptor(
 
 	return &serviceDisruptor{
 		ctx:        ctx,
-		name:       name,
+		service:    service,
 		namespace:  namespace,
 		options:    options,
-		k8s:        k8s,
-		service:    service,
-		endpoints:  ep,
-		mapper:     mapper,
 		controller: controller,
 	}, nil
-}
-
-// getTargetPods collects the name of all the target pods
-func getTargetPods(ep *corev1.Endpoints) []string {
-	var targets []string
-	for _, subset := range ep.Subsets {
-		for _, addr := range subset.Addresses {
-			if addr.TargetRef.Kind == "Pod" {
-				targets = append(targets, addr.TargetRef.Name)
-			}
-		}
-	}
-
-	return targets
-}
-
-// PortMapper maps ports of a service to a list of pod, port pairs
-type PortMapper struct {
-	ctx       context.Context
-	service   *corev1.Service
-	endpoints *corev1.Endpoints
-}
-
-// getTargetPort returns the service port that corresponds to the given port number
-// if the given port is 0, it will return the default port or error if more than one port is defined
-func (m *PortMapper) getTargetPort(port uint) (corev1.ServicePort, error) {
-	ports := m.service.Spec.Ports
-	if port != 0 {
-		for _, p := range ports {
-			if uint(p.Port) == port {
-				return p, nil
-			}
-		}
-		return corev1.ServicePort{}, fmt.Errorf("the service does not expose the given port: %d", port)
-	}
-
-	if len(ports) > 1 {
-		return corev1.ServicePort{}, fmt.Errorf("service exposes multiple ports. Port option must be defined")
-	}
-
-	return ports[0], nil
-}
-
-// Map returns for a port in a service, the EndpointPort and the list pod, port pairs
-func (m *PortMapper) Map(port uint) (map[string]uint, error) {
-	targets := map[string]uint{}
-	tp, err := m.getTargetPort(port)
-	if err != nil {
-		return targets, err
-	}
-
-	// iterate over sub-ranges looking for those that have the target port
-	// and retrieve the name of the pods and the target por
-	for _, subset := range m.endpoints.Subsets {
-		for _, p := range subset.Ports {
-			if p.Name == tp.Name {
-				for _, addr := range subset.Addresses {
-					if addr.TargetRef.Kind == "Pod" {
-						targets[addr.TargetRef.Name] = uint(p.Port)
-					}
-				}
-				break
-			}
-		}
-	}
-
-	return targets, nil
-}
-
-// NewPortMapper creates a new port mapper for a service
-func NewPortMapper(
-	ctx context.Context,
-	service *corev1.Service,
-	endpoints *corev1.Endpoints,
-) PortMapper {
-	return PortMapper{
-		ctx:       ctx,
-		service:   service,
-		endpoints: endpoints,
-	}
 }
 
 func (d *serviceDisruptor) InjectHTTPFaults(
@@ -174,7 +78,7 @@ func (d *serviceDisruptor) InjectHTTPFaults(
 	duration time.Duration,
 	options HTTPDisruptionOptions,
 ) error {
-	targets, err := d.mapper.Map(fault.Port)
+	targets, err := d.helper.MapPort(d.ctx, d.service, fault.Port)
 	if err != nil {
 		return fmt.Errorf("error getting target for fault injection: %w", err)
 	}
@@ -197,7 +101,7 @@ func (d *serviceDisruptor) InjectGrpcFaults(
 	duration time.Duration,
 	options GrpcDisruptionOptions,
 ) error {
-	targets, err := d.mapper.Map(fault.Port)
+	targets, err := d.helper.MapPort(d.ctx, d.service, fault.Port)
 	if err != nil {
 		return fmt.Errorf("error getting target for fault injection: %w", err)
 	}
